refactor: Move wheel exceptions to their submodule
